Update konec.js
Locked, hope so.

+++ commands/konec.js
@@ -6,6 +6,9 @@ const { promptMessage } = require("../functions.js");
 module.exports.run = async (bot, message, args) => {
         const logChannel = message.guild.channels.find(c => c.name === "🥋tréninky") || message.channel;
 
+        const logChan = message.guild.channels.find(c => c.name === "🌳tréninkový-ostrov");
+    if(message.channel != logChan) return message.member.send("**Konec tréninku je povolen jen zde:** <#678514400021315595>");
+
         if (message.deletable) message.delete();
         
 
@@ -38,32 +41,45 @@ module.exports.run = async (bot, message, args) => {
         logChannel.send("<@350294185972334603>, tvůj student právě dokončil trénink! Tady je zápis toho nejdůležitěšího:");
         logChannel.send(embed)
                      
-        message.channel.send(promptEmbed)
+        logChan.send(promptEmbed)
         .then(m => m.delete(30000));       
+        message.channel.send(args.join(" "));
     }
 
     if (message.member.roles.has("671814447395766273")) { //Země
         logChannel.send("<@379727594179919873>, tvůj student právě dokončil trénink! Tady je zápis toho nejdůležitějšího:");
         logChannel.send(embed)
                      
-        message.channel.send(promptEmbed)
+        logChan.send(promptEmbed)
         .then(m => m.delete(30000));  
+        message.channel.send(args.join(" "));
     }
 
     if (message.member.roles.has("671814292911423551")) { //Oheň
         logChannel.send("<@372759997211344905>, tvůj student právě dokončil trénink! Tady je zápis toho nejdůležitějšího:");
         logChannel.send(embed)
                      
-        message.channel.send(promptEmbed)
+        logChan.send(promptEmbed)
         .then(m => m.delete(30000));  
+        message.channel.send(args.join(" "));
     }
 
     if (message.member.roles.has("671814391896735770")) { //Vzduch
         logChannel.send("<@238725481153691650>, tvůj student právě dokončil trénink! Tady je zápis toho nejdůležitějšího:");
         logChannel.send(embed)
                      
-        message.channel.send(promptEmbed)
+        logChan.send(promptEmbed)
         .then(m => m.delete(30000));  
+        message.channel.send(args.join(" "));
+    }
+
+    if (message.member.roles.has("696378535069220876")) { //TEST
+        logChannel.send("<@244439995799175169>, tvůj student právě dokončil trénink! Tady je zápis toho nejdůležitějšího:");
+        logChannel.send(embed)
+                     
+        logChan.send(promptEmbed)
+        .then(m => m.delete(30000))
+        logChannel.send(args.join(" "));
     }
